Extract canonical turn record model
